requirements.txt
Flask==2.0.3
google-cloud-firestore==2.3.4
google-cloud-storage==2.1.0
google-cloud-logging==3.0.0
google-cloud-error-reporting==1.5.0
gunicorn==20.1.0
six==1.16.0

Werkzeug~=2.0.3
requests~=2.27.1